requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
datasets>=2.14.0
huggingface-hub>=0.17.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.0.0
seaborn>=0.12.0
matplotlib>=3.5.0
pandas>=1.5.0
PyQt6>=6.4.0
requests>=2.28.0 